fix: gpt4all utils not emitting docs
