+++ gpt4all-bindings/typescript/package.json
@@ -10,8 +10,8 @@
     "install": "node-gyp-build",
     "prebuild": "node scripts/prebuild.js",
     "predocs:build": "node scripts/docs.js",
-    "docs:build": "documentation readme ./src/gpt4all.d.ts --parse-extension d.ts --format md --section documentation --readme-file docs/api.md",
-    "postdocs:build": "node scripts/docs.js"
+    "docs:build": "documentation readme ./src/gpt4all.d.ts --parse-extension js d.ts --format md --section documentation --readme-file ../python/docs/gpt4all_typescript.md",
+    "postdocs:buildi": "node scripts/docs.js"
   },
   "dependencies": {
     "md5-file": "^5.0.0",
